requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.0
flask==3.0.0
frozenlist==1.4.1
googletrans-py==4.0.0
google-generativeai==0.7.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==2.10
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
libretranslatepy==2.1.1
lxml==4.9.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.6.1
Pillow==10.1.0
protobuf==4.25.1
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.3
pydantic-core==2.14.6
pyseto==1.7.7
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tabulate==0.9.0
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
webencodings==0.5.1
werkzeug==3.0.1
yarl==1.9.4